Refactoring (created BaseActivity)
