+++ app/src/main/AndroidManifest.xml
@@ -16,9 +16,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ResultActivity" />
-        <activity android:name=".SettingsActivity"></activity>
-        <activity android:name=".ReadSectionActivity"></activity>
+        <activity android:name=".BaseActivity"></activity>
     </application>
 
 </manifest>
