+++ app/src/main/res/layout/activity_base.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/result_drawer_layout"
+    android:id="@+id/base_activity_drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -17,7 +17,7 @@
     </RelativeLayout>
 
     <FrameLayout
-        android:id="@+id/result_fragment_container"
+        android:id="@+id/base_fragment_container"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:orientation="horizontal">
@@ -25,7 +25,7 @@
     </FrameLayout>
 
     <android.support.design.widget.NavigationView
-        android:id="@+id/result_navigationView"
+        android:id="@+id/base_navigationView"
         style="@style/MainNavigationDrawerDimens"
         app:headerLayout="@layout/navigation_header_main"
         app:menu="@menu/menu_navigation">
