--- app/src/main/res/layout/activity_read_section.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/read_section_drawer_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <include
-            layout="@layout/toolbar_main"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-
-    </RelativeLayout>
-
-    <FrameLayout
-        android:id="@+id/read_section_fragment_container"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-    </FrameLayout>
-
-    <android.support.design.widget.NavigationView
-        android:id="@+id/read_section_navigationView"
-        style="@style/MainNavigationDrawerDimens"
-        app:headerLayout="@layout/navigation_header_main"
-        app:menu="@menu/menu_navigation">
-
-    </android.support.design.widget.NavigationView>
-
-</android.support.v4.widget.DrawerLayout>
